Show listings near user on map

+++ Listed Near Me/Models/Location.swift
@@ -9,6 +9,6 @@
 import Foundation
 
 struct Location: Codable {
-    var long: Float
-    var lat: Float
+    var long: Double
+    var lat: Double
 }
